+++ serviceTest/CHANGELOG.md
@@ -4,10 +4,14 @@ This file documents changes to the `workbench-service-test` library, including n
 
 ## 4.3
 
-SBT dependency: `"org.broadinstitute.dsde.workbench" %% "workbench-service-test" % "4.3-bce5c47"`
+SBT dependency: `"org.broadinstitute.dsde.workbench" %% "workbench-service-test" % "4.3-TRAVIS-REPLACE-ME"`
 
 ### Changed
 - `CleanUp.runCodeWithCleanup` no longer throws an exception when the cleanup portion fails
+
+| Dependency   |     Old Version      |          New Version |
+|----------|:--------------------:|---------------------:|
+| jose4j      |    0.9.3    |       0.9.4 |
 
 ## 4.2